requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.10.2
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==22.2.0
billiard==3.6.4.0
cachetools==5.3.0
catboost==1.1.1
celery==5.2.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
configobj==5.0.8
contourpy==1.0.7
-e git+https://github.com/pk1308/creditcard.git@a800d4b6a92b3a359a2a2069e56effcfc9137cdd#egg=CreditCard
cryptography==40.0.1
cycler==0.11.0
dictdiffer==0.9.0
dill==0.3.6
diskcache==5.4.0
distro==1.8.0
dnspython==2.3.0
dpath==2.1.5
dulwich==0.21.3
dvc==2.47.2
dvc-data==0.42.3
dvc-gdrive==2.19.1
dvc-http==2.30.2
dvc-objects==0.21.1
dvc-render==0.3.1
dvc-studio-client==0.6.1
dvc-task==0.2.0
ensure==1.0.2
evidently==0.2.8
fastapi==0.95.0
filelock==3.10.7
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
funcy==2.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
grandalf==0.8
graphviz==0.20.1
greenlet==2.0.2
h11==0.14.0
httplib2==0.22.0
hydra-core==1.3.2
idna==3.4
importlib-resources==5.12.0
iterative-telemetry==0.0.8
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
lightgbm==3.3.5
loguru==0.6.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
multidict==6.0.4
nanotime==0.5.2
networkx==3.0
nltk==3.8.1
numpy==1.23.5
oauth2client==4.1.3
omegaconf==2.3.0
optuna==3.1.0
orjson==3.8.9
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
patsy==0.5.3
Pillow==9.4.0
platformdirs==3.2.0
plotly==5.14.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
pydot==1.4.2
PyDrive2==1.15.1
pygit2==1.11.1
Pygments==2.14.0
pygtrie==2.5.0
pymongo==4.3.3
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-box==6.0.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
regex==2023.3.23
requests==2.28.2
rich==13.3.3
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
scmrepo==0.1.17
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.7
sqltrie==0.3.0
starlette==0.26.1
statsmodels==0.13.5
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tomlkit==0.11.7
tqdm==4.65.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
yarl==1.8.2
yellowbrick==1.5
zc.lockfile==3.0.post1
zipp==3.15.0
gradio